Build python 3.5 async tests and keep python 2.7 tests with conditional check; Rename files to `test_*.py` to match latest test file conventions.
